Use Module.prepend when possible
This change allows both patches in this gem to utilize Ruby 2's
`Module.prepend` instead of ActiveSupport's `alias_method_chain`
when possible.

Normally I would drop support of Ruby < 2 and just
rely on prepend, but as this gem is largely for legacy purposes
this now allows for both methods depending on the version of Ruby used.

This removes current deprecation warnings with Rails 5 and will allow
continued use of the gem once alias_method_chain is removed from
ActiveSupport.